src/modules/users/services/AuthenticateUser.service.ts
import { sign } from 'jsonwebtoken';
import { injectable, inject } from 'tsyringe';

import AppError from '@shared/errors/AppError';
import authConfig from '@config/auth';
import User from '@modules/users/infra/typeorm/entities/User';

import IHashProvider from '@modules/users/providers/HashProvider/models/IHashProvider';

import IUsersRepository from '@modules/users/repositories/IUserRepository';

interface IRequest {
  email: string;
  password: string;
}

interface IResponse {
  user: User;
  token: string;
}

@injectable()
class AuthenticateUserService {
  constructor(
    @inject('UsersRepository')
    private usersRepository: IUsersRepository,

    @inject('HashProvider')
    private hashProvider: IHashProvider,
  ) {}

  public async execute({ email, password }: IRequest): Promise<IResponse> {
    const user = await this.usersRepository.findByEmail(email);
    if (!user) {
      throw new AppError('incorrect email/password .', 401);
    }

    const match = await this.hashProvider.compareHash(password, user.password);

    if (!match) {
      throw new AppError('incorrect email/password .', 401);
    }

    const { expiresIn, secret } = authConfig.jwt;

    console.log(expiresIn, secret, user);

    const token = sign({}, secret, {
      subject: `${user.id}`,
      expiresIn,
    });

    return { user, token };
  }
}

export default AuthenticateUserService;
